changed rt60 type hint

+++ src/spidam/model.py
@@ -71,7 +71,7 @@ class Model:
         return (self._frequencies[low_mask], self._frequencies[mid_mask], self._frequencies[high_mask])
             
     @functools.cache
-    def calculate_rt60(self, low_cutoff: int = 60, low_max: int = 250, mid_max: int = 5000, high_cutoff: int = 10000, decay_db: int = 60) -> Tuple[int, int, int]:
+    def calculate_rt60(self, low_cutoff: int = 60, low_max: int = 250, mid_max: int = 5000, high_cutoff: int = 10000, decay_db: int = 60) -> Tuple[float, float, float]:
         """ Generate a tuple containing the low, mid, high rt60
         """
         filtered_frequencies: Tuple[NDArray, NDArray, NDArray] = self.get_frequencies(low_cutoff, low_max, mid_max, high_cutoff)
